Dynamically sized lines and links.
Previously, lines and links had fixed length buffers.  For small lines
and links, that was wasteful.  For extremely long lines and links, that
was responsible for crashes.

Now, the structs for both line and link structs have flexible array
members to enable variable length contents.  Since the link struct has
two variable length strings, its flexible array member is used to store
both.  To facilitate accessing the two strings, the struct now has two
functions associated with it linkGetTitle and linkGetURL.

Since the size of the contents of the line and link structs must now be
known before allocating the structs which store them, the code which
constructs each had to be reordered to determine the contents first and
then construct the struct instances with those contents.

Similarly, the parsing functions have been modified to use a single
variable sized array for the line, enabled by using getline rather than
fgets.  This single buffer is used not just by parseTXT but also by
handleMarkdownStyleLink which no longer copies the title and url from
the line but instead uses pointers into the buffer and lengths in order
to construct the mdlink struct.